# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "4.0.0",
        "title": "MAGE (Mobile Awareness GEOINT Environment)",
        "description": "MAGE API",
        "contact": {
            "name": "Benjamin Tuttle",
            "email": "benjamin.t.tuttle@nga.mil"
        },
        "license": {
            "name": "Apache2",
            "url": "http://www.apache.org/licenses/LICENSE-2.0"
        }
    },
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/api/roles/{id}": {
            "get": {
                "description": "get role by id",
                "operationId": "getRoleById",
                "tags": [
                    "Role"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of role to fetch",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "role response",
                        "schema": {
                            "$ref": "#/definitions/Role"
                        }
                    }
                }
            }
        },
        "/api/devices/{id}": {
            "get": {
                "description": "Returns a single device based on id",
                "operationId": "getDeviceById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of device to fetch",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Device"
                ],
                "responses": {
                    "200": {
                        "description": "device response",
                        "schema": {
                            "$ref": "#/definitions/Device"
                        }
                    }
                }
            },
            "put": {
                "description": "update a single device based on id",
                "operationId": "updateDeviceById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of device to update",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "device update body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/DeviceCreate"
                        }
                    }
                ],
                "tags": [
                    "Device"
                ],
                "responses": {
                    "200": {
                        "description": "device update response",
                        "schema": {
                            "$ref": "#/definitions/Device"
                        }
                    }
                }
            },
            "delete": {
                "description": "deletes a single device based on id",
                "operationId": "deleteDevice",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of device to delete",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Device"
                ],
                "responses": {
                    "204": {
                        "description": "user deleted"
                    }
                }
            }
        },
...
        "/api/teams/{id}": {
            "get": {
                "description": "Returns a single team based on id",
                "operationId": "getTeamById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of team to fetch",
                        "required": true,
                        "type": "number"
                    }
                ],
                "tags": [
                    "Team"
                ],
                "responses": {
                    "200": {
                        "description": "team response",
                        "schema": {
                            "$ref": "#/definitions/Team"
                        }
                    }
                }
            },
            "put": {
                "description": "update a single team based on id",
                "operationId": "updateTeamById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of team to update",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "team update body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/TeamCreate"
                        }
                    }
                ],
                "tags": [
                    "Team"
                ],
                "responses": {
                    "200": {
                        "description": "team update response",
                        "schema": {
                            "$ref": "#/definitions/Team"
                        }
                    }
                }
            },
            "delete": {
                "description": "deletes a single team based on id",
                "operationId": "deleteTeam",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of team to delete",
                        "required": true,
                        "type": "number"
                    }
                ],
                "tags": [
                    "Team"
                ],
                "responses": {
                    "204": {
                        "description": "team deleted"
                    }
                }
            }
        },
...
        "/api/sensorservers": {
            "get": {
                "description": "get all sensor servers",
                "operationId": "getSensorServers",
                "tags": [
                    "SensorServer"
                ],
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "type": "string",
                        "description": "layer type filter, i.e. Imagery or Feature",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "sensor server response",
                        "schema": {
                            "$ref": "#/definitions/SensorServer"
                        }
                    }
                }
            },
            "post": {
                "description": "add new sensor server",
                "operationId": "addSensorServer",
                "tags": [
                    "SensorServer"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "new sensor server body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SensorServerCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "new sensor server response",
                        "schema": {
                            "$ref": "#/definitions/SensorServer"
                        }
                    }
                }
            }
        },
        "/api/sensorservers/{id}": {
            "get": {
                "description": "Returns a single sensor server based on id",
                "operationId": "getSensorServerById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of sensor server to fetch",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "SensorServer"
                ],
                "responses": {
                    "200": {
                        "description": "sensor server response",
                        "schema": {
                            "$ref": "#/definitions/SensorServer"
                        }
                    }
                }
            },
            "put": {
                "description": "update a single sensor server based on id",
                "operationId": "updateSensorServerByName",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of sensor server to update",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "sensor server update body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SensorServerCreate"
                        }
                    }
                ],
                "tags": [
                    "SensorServer"
                ],
                "responses": {
                    "200": {
                        "description": "sensor server update response",
                        "schema": {
                            "$ref": "#/definitions/SensorServer"
                        }
                    }
                }
            },
            "delete": {
                "description": "deletes a single sensor server based on id",
                "operationId": "deleteSensorServer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of sensor server to delete",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "SensorServer"
                ],
                "responses": {
                    "204": {
                        "description": "sensor server deleted"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Role": {
            "required": [
                "id",
                "name",
                "description",
                "permissions"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "User": {
            "required": [
                "id",
                "username",
                "displayName",
                "active",
                "role"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "active": {
                    "type": "boolean"
                },
                "role": {
                    "$ref": "#/definitions/Role"
                },
                "avatarUrl": {
                    "type": "string"
                },
                "iconUrl": {
                    "type": "string"
                }
            }
        },
...
        "DeviceCreate": {
            "properties": {
                "uid": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "poc": {
                    "type": "string"
                },
                "registered": {
                    "type": "boolean",
                    "description": "boolean indicating if device is registered, ADMIN only"
                }
            }
        },
        "Device": {
            "required": [
                "id",
                "uid",
                "name",
                "description"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "registered": {
                    "type": "boolean"
                }
            }
        },
        "TeamCreate": {
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/User"
                    }
                }
            }
        },
        "Team": {
            "required": [
                "id",
                "name"
            ],
            "properties": {
                "id": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/User"
                    }
                }
            }
        },
...
        "SensorServerCreate": {
            "required": [
                "name",
                "url"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "SensorServer": {
            "required": [
                "id",
                "name",
                "url"
            ],
            "properties": {
                "id": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                }
            }
        },
...
    }
}